requirements.txt
# Core dependencies
notion-client>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# CLI and UI
rich>=13.0.0
inquirer>=3.0.0

# LLM providers (optional - install as needed)
anthropic>=0.20.0
openai>=1.0.0
google-generativeai>=0.3.0